re #50
I've fixed a couple of bugs related to removing a favorite. 
1) It was possible to remove a favorite right after favoriting it, but it was not possible to do it if you went to another page and came back later to the sentence. 
2) Removing a favorite was also deleting ALL the favorites instead (which we obviously don't want)... I'm not sure why, but anyway I've redefined the Favorite model so that the HABTM association is User and not Favorite.

There was also a problem while favoriting from the search page because several sentences are displayed. I've fixed that.

Instead of displaying the loading gif for the whole set of sentences, I'm dislaying a smaller gif next to the (un)favorite link.

As for testing, I've tried favoriting a sentence with two different users. Deleting from a user doesn't seem to delete the favorite for the other user. I think it's pretty much okay now. 

@sysko, I'll wait for you to try it out and close this ticket if there's no more major bug and if you're okay with how it works. Then I'll update on the production server so you can finally have this feature you were waiting for so long :P

Note : one thing that would be practical is to enable the user to unfavorite a sentence directly from the page listing the favorites.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@308 e0e46c49-be69-4f5a-ad62-21024a331aea