re #82. My contribution for the design.
Header is now vertical (on the left). It saves us some pixels on the height.

I decided to put the search bar on top of the main content instead of putting it on the right corner, and I made it adaptable to the size of the main content (which was a real pain by the way... and it's still not perfect but I don't want to work on it any more). This way it doesn't look too weird on big resolutions compared to when it was it on the right corner (I tried 1680x1050).

There's still a problem : the min-width. That is, if you reduce too much the size of the browser, it will start breaking the design.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@335 e0e46c49-be69-4f5a-ad62-21024a331aea