re #51.
The list selection appears after clicking on "Add to list". You can hide it by re-clicking on "Add to list" (it's a toggle type of thing).

If a sentence already belongs to a list, that list won't appear in the <select>. Following that idea, I made it so that after you have added a sentence to a list, the list name is removed from the <select>.

I've also added a "Manage lists" that redirects to the lists page.

I'm still not fully satisfied with the feature. When I select a list, sometimes I kind of want to see what are the sentences in that list. Also, the feedback after adding to a list is not great... it doesn't feel like the sentence has actually been added.

And I still have to get the feature work properly on the search results.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@344 e0e46c49-be69-4f5a-ad62-21024a331aea