re #82.
Added info icon for sentences and translations. Clicking on the icon displays the sentence in the Browse section.
When the text of the translations was a link, it was a bit annoying to highlight a specific word of the translation. That's why I'm using an icon instead. Hopefully it's obvious enough that the icon is clickable...

Improved display for sentences comments.

In the process of improving the contribute page.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@380 e0e46c49-be69-4f5a-ad62-21024a331aea