re #99.
The key was to add these two lines :
echo $javascript->link('jquery.jeditable.js', true);
echo $javascript->link('sentences.edit_in_place.js', true);

Remember: when you load content in AJAX (with jQuery at least?), the part that is loaded does not inherit of the javascript of the page. You need to explicitely call the javascripts again, within the loaded content.

(I don't know why but if you have an explaination for this, I'd be glad to know)

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@521 e0e46c49-be69-4f5a-ad62-21024a331aea