Following Swift's request : http://tatoeba.org/eng/wall/index#message_85
Took out info and comment icon in sentence menu. 
Made the sentence clickable instead, except in Browse section where it can be editable if the sentence belongs to current user.
Took out the hidden info button next to the translations. Instead there's an arrow icon (it's a different one whether the translation is direct or indirect). The icon is clickable.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@605 e0e46c49-be69-4f5a-ad62-21024a331aea