re #103.
Took out the "same language" warning message. 
First of all, it was there was a bug I couldn't figure out (the warning message would be displayed several times). 
Second of all, considering that we're hiding the translations, people who are going to add a sentence of the same language will know what they're doing.
And last but not least, it makes the code a bit clearer now because the save_translation/check_translation was a bit confusing.

It could be a bit annoying to see the "Important!" message every time, but I don't think users will mind. If we wanted to enable users to hide this message, it would require modifying the database to save who doesn't want to see the message, and I don't want to go through this trouble for something small.

As for being able to submit with "enter", it should be possible... At least I can do it...

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@803 e0e46c49-be69-4f5a-ad62-21024a331aea