re #163.
Alright today I learned: sometimes you have to unbind handlers. The problem was that everytime you click on the translate icon, it binds the handlers for submitting the translation. So the translation would be saved as many times as you would have clicked on the translate icon.

I hope this solves the problem of being automatically disconnected when not checking the "remember me" and adding two translations in a row for the same sentence. At least it works in my local version.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@844 e0e46c49-be69-4f5a-ad62-21024a331aea