test #104
It was again a problem with a handler being binded several times. As I have written in the comments of the Javascript:

It's important to unbind because when adding two translations in a row, it will rebind again the same function. So the second time the user clicks on a flag, the function is triggered twice and therefore it looks like nothing happens. But what happens is that the <select> is displayed then hidden right away.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@853 e0e46c49-be69-4f5a-ad62-21024a331aea