re #188.
Fixed problem with search bar on Opera. This took me forever but the search bar is now properly display on the latest versions of Firefox, Chrome, Opera and Internet Explorer. Hopefully it works on Safari too =_='

Corrected class of links in elements/space.ctp so that they appear in white, not in greyish.

And other small things...

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@924 e0e46c49-be69-4f5a-ad62-21024a331aea