Refactoring.
Now using CurrentUser::canLinkAndUnlink($sentenceId) in save_translation() method of sentences_controller.php in order to tell whether a user can link/unlink.

I think we will move all the permissions stuff into the CurrentUser class. It's a singleton which enables you to access the "Auth" information from anywhere, instead of only from controllers. It was first used to be able to retrieve the current user id in the ContributionModel.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1006 e0e46c49-be69-4f5a-ad62-21024a331aea