re #220.
I didn't really added a log table. Instead I added two fields in the "sentence_annotations" table: 'modified' and 'user_id'. Each time a user modifies an entry, it updates the 'modified' field and the 'user_id' fields accordingly. 

It doesn't enable to keep track of the the evolution of an annotation, but Jim and Paul do not really need it I think. They mostly need to know what were the last modifications made on the indices.

They may however be interested in knowing about the entries that have been deleted... But this is not extremely urgent, I think. The annotations are restricted to trusted user so there isn't much of a risk of vandalism or anything.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1141 e0e46c49-be69-4f5a-ad62-21024a331aea