re #216.
Took out isFavorited from SentenceModel. Instead, we now check whether a sentence is favorited by the current user or not from the CurrentUserModel. 

To do so, I changed the value $useTable in FavoritedBy from 'users' into 'favorites_users'. Then created a method isSentenceFavoritedByUser(), and used this method from the CurrentUserModel.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1145 e0e46c49-be69-4f5a-ad62-21024a331aea