test #204.
Added info icon when sentence is editable. (I had to do this one... as the admin, ALL the sentences are editable to me, can you imagine the pain...).

NOTE: in terms of UI, it is NOT good to have the icon appear while in "Browse". It is NOT good either to have the sentence as a link while in "Browse". But it's a minor UI problem.

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1162 e0e46c49-be69-4f5a-ad62-21024a331aea