test #236.
Using SentencesHelper::displaySentencesGroup() to display sentences in list. Now you can do everything with sentences in list.
It makes the list a too crowded in my opinion, but I will try to improve this some other time.

It would also be nice to have pagination, because displaying 500 sentences like this is not optimal...

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1199 e0e46c49-be69-4f5a-ad62-21024a331aea