re #271.
- generic_functions.js
- sentences.add_translation.js

Added function to remove the new lines and double (or more) spaces. This is mostly to remove new lines and tabs though. 

The only place where new lines can be inserted is when people add translations, because the translation form is actually a textarea. If it was an input, new lines would be automatically removed when people copy-paste but we want to keep it as a textarea for spellcheck and because it gives the possibility to see the whole sentence when it's a long sentence.

Anyway, removing the new lines fromt he javascript should avoid (for now) having to check for new lines and remove them in the cleaner-script-that-doesnt-exist-yet (but there should still be something for spaces).

git-svn-id: https://subversion.assembla.com/svn/tatoeba2/trunk@1661 e0e46c49-be69-4f5a-ad62-21024a331aea